+++ src/components/Carousel.tsx
@@ -3,6 +3,7 @@ import kablokanal from "../assets/images/slides/kablokanal.png";
 import kablolama from "../assets/images/slides/kablolama.png";
 import kablorekor from "../assets/images/slides/kablorekor.png";
 import trenkantar from "../assets/images/slides/trenkantar.png";
+import musterikablo from "../assets/images/slides/müsterikabloc.png";
 import "../styles/Carousel.css";
 
 const CarouselExample = () => {
@@ -45,6 +46,22 @@ const CarouselExample = () => {
         </div>
         <Carousel.Caption>
           <h3>Endüstriyel markalama ve kablo bağları</h3>
+        </Carousel.Caption>
+      </Carousel.Item>
+
+      <Carousel.Item interval={3000} >
+        <div className="carousel-image-container">
+          <img
+            src={musterikablo}
+            className="carousel-image custom-width"
+            alt="musterikablo"
+          />
+        </div>
+        <Carousel.Caption>
+          <h3 className="bg-dark bg-opacity-50 p-2">
+            Müşteriye Özel Kablolu Çözümler
+          </h3>
+          <p>Fiber Cords, Patch Cords, Jumper Cables</p>
         </Carousel.Caption>
       </Carousel.Item>
     </Carousel>